solves issue #4. Compatible max parameter types no longer deleted. Function AreTypesCompatible prototyped, defined and used to test for compatibility.

+++ src/FabricTranslationLayer.hpp
@@ -769,9 +769,13 @@ int FabricTranslationLayer<TBaseClass, TResultType>::SyncMetaDataFromPortToParam
         // Check that its not the correct type already
         if (m_pblock->GetParamDef((ParamID)paramId).type != maxType)
         {
-          // Is the type incompatible?  For now... guess so...
-          DeleteMaxParameter((ParamID)paramId);
-          paramId = -1;
+			
+			// Check that the types are compatible
+			if(!AreTypesCompatible(m_pblock->GetParamDef((ParamID)paramId).type, maxType)) 
+			{
+				DeleteMaxParameter((ParamID)paramId);
+				paramId = -1;
+			}
         }
       }
     }
